+++ packages/elements/src/__stories__/components/StoplightProject.tsx
@@ -7,9 +7,9 @@ import * as React from 'react';
 import { StoplightProject } from '../../containers/StoplightProject';
 
 const darkMode = () => boolean('dark mode', false);
-const workspace = () => text('workspace', 'https://elements.stoplight.io');
-const project = () => text('project', 'public-apis');
-const branch = () => text('branch', '');
+const workspaceSlug = () => text('workspaceSlug', 'elements');
+const projectSlug = () => text('projectSlug', 'public-apis');
+const branchSlug = () => text('branchSlug', '');
 const authToken = () => text('authToken', '');
 
 storiesOf('Public/StoplightProject', module)
@@ -17,7 +17,12 @@ storiesOf('Public/StoplightProject', module)
   .add('Playground', () => {
     return (
       <div className={cn('p-10', { 'bp3-dark bg-gray-8': darkMode() })}>
-        <StoplightProject workspace={workspace()} project={project()} branch={branch()} authToken={authToken()} />
+        <StoplightProject
+          workspaceSlug={workspaceSlug()}
+          projectSlug={projectSlug()}
+          branchSlug={branchSlug()}
+          authToken={authToken()}
+        />
       </div>
     );
   });