+++ packages/elements/src/__stories__/index.ts
@@ -2,7 +2,6 @@
 import './components/Api';
 import './components/StoplightProject';
 // Internal Components
-import './components/Dependencies';
 import './components/Docs';
 import './components/EditMode';
 import './components/TableOfContents';