+++ packages/elements/src/components/Docs/HttpOperation/Parameters.tsx
@@ -5,7 +5,7 @@ import cn from 'classnames';
 import { capitalize, get, isEmpty, keys, omit, omitBy, pick, pickBy, sortBy } from 'lodash';
 import * as React from 'react';
 
-import { InlineRefResolverContext } from '../../../containers/Provider';
+import { useInlineRefResolver } from '../../../context/InlineRefResolver';
 import { useClasses } from '../../../hooks/useClasses';
 import { useClick } from '../../../hooks/useClick';
 import { useStyle } from '../../../hooks/useStyle';
@@ -58,7 +58,7 @@ export const Parameters: React.FunctionComponent<IParametersProps> = ({
   className,
   onClick,
 }) => {
-  const resolveRef = React.useContext(InlineRefResolverContext);
+  const resolveRef = useInlineRefResolver();
   if (!parameters || !parameters.length) return null;
 
   return (
@@ -217,7 +217,7 @@ const NumberValidations = ({ validations, className }: { validations: Dictionary
       const sign = `${key.includes('min') ? '>' : '<'}${exclusive ? '' : '='}`;
 
       return (
-        <div key={key} className={cn('ml-2 text-sm bp3-running-text', className)}>
+        <div key={key} className={cn('ml-2 text-sm bp3-running-text break-all', className)}>
           <code>{`${sign} ${validations[key]}${suffix}`}</code>
         </div>
       );
@@ -253,7 +253,7 @@ const KeyValueValidation = ({
   }
   const validation = Array.isArray(value) ? value : [value];
   return (
-    <div className={cn('text-sm mt-2 bp3-running-text', className)}>
+    <div className={cn('text-sm mt-2 bp3-running-text break-all flex flex-wrap', className)}>
       {capitalize(name)}:
       {validation
         .filter(