+++ packages/elements/src/constants.ts
@@ -9,6 +9,8 @@ import {
 } from '@fortawesome/free-solid-svg-icons';
 import { Dictionary, HttpSecurityScheme, NodeType } from '@stoplight/types';
 
+export const defaultPlatformUrl = 'https://stoplight.io';
+
 export const NodeTypeColors: Dictionary<string, NodeType> = {
   http_operation: '#6a6acb',
   http_service: '#e056fd',
@@ -40,17 +42,6 @@ export const NodeTypeIconDefs: Dictionary<IconDefinition, NodeType> = {
   unknown: faQuestionCircle,
   generic: faQuestionCircle,
   table_of_contents: faQuestionCircle,
-};
-
-export const NodeTypeIconsUnicode: Dictionary<string, NodeType> = {
-  http_operation: '\uf140',
-  http_service: '\uf0c2',
-  article: '\uf02d',
-  model: '\uf1b2',
-  http_server: '\uf1c0',
-  unknown: '\uf128',
-  generic: '\uf128',
-  table_of_contents: '\uf128',
 };
 
 export const HttpSecuritySchemeColors: Dictionary<string, HttpSecurityScheme['type']> = {