+++ packages/elements/src/containers/API.tsx
@@ -10,6 +10,7 @@ import useSwr from 'swr';
 import { SidebarLayout } from '../components/API/SidebarLayout';
 import { StackedLayout } from '../components/API/StackedLayout';
 import { DocsSkeleton } from '../components/Docs/Skeleton';
+import { InlineRefResolverProvider } from '../context/InlineRefResolver';
 import { withRouter } from '../hoc/withRouter';
 import { useBundleRefsIntoDocument } from '../hooks/useBundleRefsIntoDocument';
 import { useParsedValue } from '../hooks/useParsedValue';
@@ -18,20 +19,34 @@ import { LinkComponentType, RoutingProps } from '../types';
 import { computeNodeData, isOas2, isOas3, IUriMap } from '../utils/oas';
 import { computeOas2UriMap } from '../utils/oas/oas2';
 import { computeOas3UriMap } from '../utils/oas/oas3';
-import { InlineRefResolverProvider } from './Provider';
 
 const fetcher = (url: string) => axios.get(url).then(res => res.data);
 
-export interface APIProps extends RoutingProps {
-  apiDescriptionUrl: string;
+export type APIProps = APIPropsWithDocument | APIPropsWithUrl;
+
+export type APIPropsWithUrl = { apiDescriptionUrl: string } & CommonAPIProps;
+export type APIPropsWithDocument = {
+  apiDescriptionDocument: string | object;
+  apiDescriptionUrl?: string;
+} & CommonAPIProps;
+
+export interface CommonAPIProps extends RoutingProps {
   linkComponent?: LinkComponentType;
   layout?: 'sidebar' | 'stacked';
 }
 
-const APIImpl = withRouter<APIProps>(function API({ apiDescriptionUrl, linkComponent, layout }) {
+const propsAreWithDocument = (props: APIProps): props is APIPropsWithDocument => {
+  return props.hasOwnProperty('apiDescriptionDocument');
+};
+
+const APIImpl = withRouter<APIProps>(function API(props) {
+  const { layout, linkComponent, apiDescriptionUrl } = props;
+  const apiDescriptionDocument = propsAreWithDocument(props) ? props.apiDescriptionDocument : undefined;
+
   const { pathname } = useLocation();
 
-  const { data, error } = useSwr(apiDescriptionUrl, fetcher);
+  const documentShouldBeFetched = apiDescriptionUrl && !apiDescriptionDocument;
+  const { data: fetchedDocument, error } = useSwr(documentShouldBeFetched ? apiDescriptionUrl! : null, fetcher);
 
   React.useEffect(() => {
     if (error) {
@@ -39,7 +54,7 @@ const APIImpl = withRouter<APIProps>(function API({ apiDescriptionUrl, linkCompo
     }
   }, [error]);
 
-  const document = useParsedValue(data);
+  const document = useParsedValue(apiDescriptionDocument || fetchedDocument);
   const bundledDocument = useBundleRefsIntoDocument(document, { baseUrl: apiDescriptionUrl });
 
   const uriMap = React.useMemo(() => {